requirements.txt
asgiref==3.2.10
atomicwrites==1.4.1
attrs==22.1.0
colorama==0.4.5
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.0.4
importlib-metadata==4.12.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.8.6
py==1.11.0
PyJWT==2.1.0
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
pytz==2020.1
sqlparse==0.3.1
tomli==2.0.1
typing_extensions==4.3.0
zipp==3.8.1
pytest-django==4.5.2
